requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
coverage==7.9.1
cryptography==45.0.4
defusedxml==0.7.1
dj-database-url==3.0.0
Django==5.2.2
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
dotenv==0.9.9
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
#psycopg2==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
